working i-mask-text-field and number-format-text-field
